# Conflicts:
#	public/mix-manifest.json

+++ app/Data/Repositories/UserRepository.php
@@ -51,11 +51,11 @@ public $model;
             if (!empty($details['profile_data'])) {
                 if($data->role_id == Role::SERVICE_PROVIDER){
                 // Todo
-                $data->business_details = app('ServiceProviderProfileRepository')->findByAttribute('user_id' , $id,false,true);                
-                if (!empty($details['provider_request_data'])) {
-                    $serviceDetailsCriteria = ['user_id' => $id];
-                    $data->service_details = app('ServiceProviderProfileRequestRepository')->findCollectionByCriteria($serviceDetailsCriteria);                
-                }
+                    $data->business_details = app('ServiceProviderProfileRepository')->findByAttribute('user_id' , $id,false,true);                
+                    if (!empty($details['provider_request_data'])) {
+                        $serviceDetailsCriteria = ['user_id' => $id];
+                        $data->service_details = app('ServiceProviderProfileRequestRepository')->findCollectionByCriteria($serviceDetailsCriteria);                
+                    }
 
                 }
             }
@@ -99,64 +99,71 @@ public $model;
     public function update(array $data = []) {
 
         $input = $data['user_details'];
-        $input['id'] = $data['user_id'];
+        $input['id'] = $data['id'];
         
         if ($user = parent::update($input)) {
 
-            if(!empty($data['business_details'])){
 
-                $business_details = $data['business_details']; 
-                $business_details['user_id'] = $user->id;
-                if($business = app('ServiceProviderProfileRepository')->findByAttribute('user_id' , $user->id)){
-                    $business_details['id'] = $business->id;
-                    $user->business_details = app('ServiceProviderProfileRepository')->update($business_details);
-                }else{
-                    $user->business_details = app('ServiceProviderProfileRepository')->create($business_details);
+            if($user->role_id == Role::SERVICE_PROVIDER){
+
+                if(!empty($data['business_details'])){
+
+                    $business_details = $data['business_details']; 
+                    $business_details['user_id'] = $user->id;
+                    if($business = app('ServiceProviderProfileRepository')->findByAttribute('user_id' , $user->id)){
+                        $business_details['id'] = $business->id;
+                        $user->business_details = app('ServiceProviderProfileRepository')->update($business_details);
+                    }else{
+                        $user->business_details = app('ServiceProviderProfileRepository')->create($business_details);
+                    }
                 }
-            }
 
-            if(!empty($data['service_details'])){
+                if(!empty($data['service_details'])){
 
-                foreach ($data['service_details'] as $key => $service) {
-                    if(empty($service['service_id'])){
+                    foreach ($data['service_details'] as $key => $service) {
+                        if(empty($service['service_id'])){
                             continue;
                         }
 
-                    if(!empty($service['id'])){
+                        if(!empty($service['id'])){
 
-                        $existingServiceIds[$service['id']][] = $service['service_id'];
-                        $service['service_provider_profile_request_id'] = $service['id'];
-                        unset($service['id']);
+                            $existingServiceIds[$service['id']][] = $service['service_id'];
+                            $service['service_provider_profile_request_id'] = $service['id'];
+                            unset($service['id']);
 
-                        $service['deleted_at'] = null;
-                        $existingServices[] = $service;
+                            $service['deleted_at'] = null;
+                            $existingServices[] = $service;
 
-                    }else{
-                        $newServices[] = $service;
+                        }else{
+                            $newServices[] = $service;
+                        }
+                    }                
+
+                    if(!empty($newServices)){
+                        $serviceProfileRequest = app('ServiceProviderProfileRequestRepository')->create(['user_id' => $user->id]);
+                        foreach ($newServices as $key => $newService) {
+                            $newServices[$key]['service_provider_profile_request_id'] = $serviceProfileRequest->id; 
+                        }
+                        app('ServiceProviderServiceRepository')->model->insert($newServices);
                     }
-                }                
 
-                if(!empty($newServices)){
-                    $serviceProfileRequest = app('ServiceProviderProfileRequestRepository')->create(['user_id' => $user->id]);
-                    foreach ($newServices as $key => $newService) {
-                        $newServices[$key]['service_provider_profile_request_id'] = $serviceProfileRequest->id; 
+                    if(!empty($existingServiceIds)){
+                        foreach ($existingServiceIds as $key => $existingServiceId) {
+                            app('ServiceProviderServiceRepository')->model
+                            ->where('service_provider_profile_request_id' , $key)
+                            ->whereNotIn('id', $existingServiceId)->delete();
+                        }
+
+                        app('ServiceProviderServiceRepository')->model->insertOnDuplicateKey($existingServices);
+
                     }
-                    app('ServiceProviderServiceRepository')->model->insert($newServices);
                 }
 
-                if(!empty($existingServiceIds)){
-                    foreach ($existingServiceIds as $key => $existingServiceId) {
-                        app('ServiceProviderServiceRepository')->model
-                        ->where('service_provider_profile_request_id' , $key)
-                        ->whereNotIn('id', $existingServiceId)->delete();
-                    }
+                $details = ['profile_data' => true , 'provider_request_data' => true];
+                $user = self::findById($user->id , true, $details);
+            }
 
-                    app('ServiceProviderServiceRepository')->model->insertOnDuplicateKey($existingServices);
-                    
-                }
 
-                $user = self::findById($user->id , true);
-            }            
             return $user;
         }
